Include Simplified BSD license
We want this to be open source but also be able to cover our asses.

Include the 2-clause BSD license mainly so that someone doesn't sue us for using
this gem in their software (no liability and no warranty clause). And that
credits gets due where credit gets due.

Could've been an MIT or Apache license but we're a California company.